+++ src/arrays/unlockableRoomTypes.ts
@@ -5,8 +5,8 @@ export const UNLOCKABLE_ROOM_TYPES = [
   RoomType.CURSE, // 10
   RoomType.LIBRARY, // 12
   RoomType.SACRIFICE, // 13
-  RoomType.CLEAN_BEDROOM, // 18
-  RoomType.DIRTY_BEDROOM, // 19
+  /// RoomType.CLEAN_BEDROOM, // 18 (uncomment if another unlock is needed)
+  /// RoomType.DIRTY_BEDROOM, // 19 (uncomment if another unlock is needed)
   RoomType.VAULT, // 20
   RoomType.DICE, // 21 (uncomment if one more unlock is needed)
   RoomType.PLANETARIUM, // 24