+++ src/constants.ts
@@ -5,3 +5,5 @@ export const MOD_NAME = "Achievement Randomizer";
 export const IS_DEV = true as boolean;
 
 export const STARTING_CHARACTER = PlayerType.ISAAC;
+
+export const LAST_VERSION_WITH_ACHIEVEMENT_CHANGES = "0.2.0";